Added config info on display. BBU1 now starts from 0

+++ include/simulate.hpp
@@ -4,14 +4,14 @@
 #include "cout.hpp"
 #include "debit.hpp"
 
-int simulate_cout();
-int full_sim_cout();
-int disp_res_cout();
+float simulate_cout();
+int   full_sim_cout();
+int   disp_res_cout();
 
 
-int simulate_debit();
-int full_sim_debit();
-int disp_res_debit();
+float simulate_debit();
+int   full_sim_debit();
+int   disp_res_debit();
 
-int save_best_Y(std::vector<float> A, char val);
+int   save_best_Y(std::vector<float> A, char val);
 #endif //SIM_HPP
